refactor(levm): decluttering vm.rs (#2733)
**Motivation**

Making the code of the vm.rs file cleaner, since it's a bit cluttered
right now.

**Description**

The code of vm.rs is a bit messy right now. This PR moves EVM config to
environment, moves a few attributes from environment to substate that
make more sense there, and removes the StateBackup struct since it's
made unnecessary by this change.

Closes #2731, Closes #2717 
Resolves most of #2718

---------
